Elevator uses a filter and properly deadbands
git-svn-id: http://svn.lynbrookrobotics.com/code15@334 9f8849d2-cf13-4b18-a094-84ffabbe6057

+++ Robot/LRT15/src/com/team846/frc2015/automation/LoadAdditional.java
@@ -18,7 +18,7 @@ public class LoadAdditional extends LoadItem implements Configurable {
     final ElevatorData elevatorData;
     private final CollectorArmData armData;
     private int toteAnalogValue = 0;
-    private int startPosition = 0;
+    private double startPosition = 0.0;
     private int hookDisengageDrop = 0;
     private boolean skipPickup = false;
